+++ _/worker/mappers/msgMapper.js
@@ -1,0 +1,5 @@
+const getCurrencies = require('../API/getCurrencies');
+
+module.exports = {
+    getCurrencies,
+}